refactor: extract handleSlash (~1,400 lines) from app.tsx
Move the slash-command dispatcher into src/app/slash-commands.ts as a
standalone function with a typed SlashCommandContext interface.

App.tsx keeps a thin useCallback wrapper (~160 lines) that builds the
context object and delegates to the extracted function.

Also removes dead state and unused imports surfaced by qlty.

Commit-Queue-Session: cq_20260508160558_dad111b7260fef0bcd71f2b9
Coding-Agent: pi
Coding-Agent-Session: pi-019e06b6-830e-72ca-8378-e2ba2d3209c0
